+++ src/wymeditor/core.js
@@ -626,28 +626,91 @@ jQuery.fn.wymeditor = function (options) {
             '</li>',
 
         toolsItems: [
-            {'name': 'Bold', 'title': 'Strong', 'css': 'wym_tools_strong'},
-            {'name': 'Italic', 'title': 'Emphasis', 'css': 'wym_tools_emphasis'},
-            {'name': 'Superscript', 'title': 'Superscript',
-                'css': 'wym_tools_superscript'},
-            {'name': 'Subscript', 'title': 'Subscript',
-                'css': 'wym_tools_subscript'},
-            {'name': 'InsertOrderedList', 'title': 'Ordered_List',
-                'css': 'wym_tools_ordered_list'},
-            {'name': 'InsertUnorderedList', 'title': 'Unordered_List',
-                'css': 'wym_tools_unordered_list'},
-            {'name': 'Indent', 'title': 'Indent', 'css': 'wym_tools_indent'},
-            {'name': 'Outdent', 'title': 'Outdent', 'css': 'wym_tools_outdent'},
-            {'name': 'Undo', 'title': 'Undo', 'css': 'wym_tools_undo'},
-            {'name': 'Redo', 'title': 'Redo', 'css': 'wym_tools_redo'},
-            {'name': 'CreateLink', 'title': 'Link', 'css': 'wym_tools_link'},
-            {'name': 'Unlink', 'title': 'Unlink', 'css': 'wym_tools_unlink'},
-            {'name': 'InsertImage', 'title': 'Image', 'css': 'wym_tools_image'},
-            {'name': 'InsertTable', 'title': 'Table', 'css': 'wym_tools_table'},
-            {'name': 'Paste', 'title': 'Paste_From_Word',
-                'css': 'wym_tools_paste'},
-            {'name': 'ToggleHtml', 'title': 'HTML', 'css': 'wym_tools_html'},
-            {'name': 'Preview', 'title': 'Preview', 'css': 'wym_tools_preview'}
+            {
+                'name': 'Bold',
+                'title': 'Strong',
+                'css': 'wym_tools_strong'
+            },
+            {
+                'name': 'Italic',
+                'title': 'Emphasis',
+                'css': 'wym_tools_emphasis'
+            },
+            {
+                'name': 'Superscript',
+                'title': 'Superscript',
+                'css': 'wym_tools_superscript'
+            },
+            {
+                'name': 'Subscript',
+                'title': 'Subscript',
+                'css': 'wym_tools_subscript'
+            },
+            {
+                'name': 'InsertOrderedList',
+                'title': 'Ordered_List',
+                'css': 'wym_tools_ordered_list'
+            },
+            {
+                'name': 'InsertUnorderedList',
+                'title': 'Unordered_List',
+                'css': 'wym_tools_unordered_list'
+            },
+            {
+                'name': 'Indent',
+                'title': 'Indent',
+                'css': 'wym_tools_indent'
+            },
+            {
+                'name': 'Outdent',
+                'title': 'Outdent',
+                'css': 'wym_tools_outdent'
+            },
+            {
+                'name': 'Undo',
+                'title': 'Undo',
+                'css': 'wym_tools_undo'
+            },
+            {
+                'name': 'Redo',
+                'title': 'Redo',
+                'css': 'wym_tools_redo'
+            },
+            {
+                'name': 'CreateLink',
+                'title': 'Link',
+                'css': 'wym_tools_link wym_opens_dialog'
+            },
+            {
+                'name': 'Unlink',
+                'title': 'Unlink',
+                'css': 'wym_tools_unlink'
+            },
+            {
+                'name': 'InsertImage',
+                'title': 'Image',
+                'css': 'wym_tools_image wym_opens_dialog'
+            },
+            {
+                'name': 'InsertTable',
+                'title': 'Table',
+                'css': 'wym_tools_table wym_opens_dialog'
+            },
+            {
+                'name': 'Paste',
+                'title': 'Paste_From_Word',
+                'css': 'wym_tools_paste wym_opens_dialog'
+            },
+            {
+                'name': 'ToggleHtml',
+                'title': 'HTML',
+                'css': 'wym_tools_html'
+            },
+            {
+                'name': 'Preview',
+                'title': 'Preview',
+                'css': 'wym_tools_preview wym_opens_dialog'
+            }
         ],
 
         containersHtml: String() +
@@ -1105,6 +1168,10 @@ WYMeditor.INIT_DIALOG = function (index) {
         sStamp = wym.uniqueStamp(),
         tableOnClick;
 
+    jQuery(window).bind('beforeunload', function () {
+        wym.focusOnDocument();
+    });
+
     if (dialogType === WYMeditor.DIALOG_LINK) {
         // ensure that we select the link to populate the fields
         if (selected && selected.tagName &&
@@ -1355,17 +1422,20 @@ WYMeditor.isPhantomString = function (str) {
 // Returns the Parents or the node itself
 // jqexpr = a jQuery expression
 jQuery.fn.parentsOrSelf = function (jqexpr) {
-    var $n = this;
+    var $n = this,
+        addBackName;
 
-    if ($n[0].nodeType === WYMeditor.NODE.TEXT) {
-        $n = $n.parents().slice(0, 1);
+    if (jQuery.fn.addBack) {
+        addBackName = 'addBack';
+    } else {
+        // `andSelf` is deprecated.
+        addBackName = 'andSelf';
     }
 
-//  if ($n.is(jqexpr)) // XXX should work, but doesn't (probably a jQuery bug)
-    if ($n.filter(jqexpr).size() === 1) {
-        return $n;
+    if (jqexpr) {
+        return $n.parents()[addBackName](jqexpr);
     } else {
-        return $n.parents(jqexpr).slice(0, 1);
+        return $n.parents()[addBackName]();
     }
 };
 
@@ -1455,5 +1525,3 @@ WYMeditor.Helper = {
         return null;
     }
 };
-
-